move interfaces to individual modules
